Fix high CPU usage by avoiding feedback loop
Depending on the size of the logo image and the
configured size it is possible, that any
_icon.set_size() call triggers a new
"allocation-changed" event which in order calls
_updateScale() and _icon.set_size() again (with
slightly different numbers).

This happens if the manually set size and the
subsequently allocated size differ a little bit
(e.g. due to rounding).

Avoid this kind of "feedback loop" by limiting the
update to cases where the current size differs
significantly from the new size.