Improve tracking of existing logos
We currently track all logos we create in an array, so we can destroy them
when no longer needed. However it is possible that the corresponding actor
has already been destroyed (for instance when a monitor is unplugged), re-
sulting in warnings with newer gjs versions.

Fix this by tracking logos in a Set and remove any logo whose actor is
destroyed.

https://pagure.io/background-logo-extension/issue/8