Provide Meta.BackgroundActor compatibility
We add ourselves to the background group, so gnome-shell will try
to animate the 'brightness' and 'vignette-sharpness' properties
when entering the overview (assuming all background actors are
Meta.BackgroundActors).

In the past this caused a harmless warning, but after dropping Tweener
for animations, we now trigger an error that messes up the overview.

Fix this by providing the same properties as Meta.BackgroundActor.

https://pagure.io/background-logo-extension/pull-request/16