prefs: Support handling slide shows without GnomeDesktop
GnomeDesktop is a dependency of gnome-shell itself, so it is safe to
assume that it is available. However extension prefs must port to GTK4
now, and gnome-desktop is still gtk3-only at this point.

In order to not let GnomeDesktop block the port, add a simplistic
fallback handling for slide shows, where we simply grep the first
<file> element from the XML.