prefs: Don't try to use actual monitor ratio
The current way for (maybe) figuring out the ratio relies on
deprecated API that is no longer available in GTK4.

The new API works properly on wayland, but is tied to surfaces
which makes it fiddlier to use. Considering that matching the
output ratio doesn't really provide any value, just show the
preview at a fixed ratio of 16:9.